requirement.txt
django<=5.0.0
djangorestframework
django-cors-headers
django-environ
django-contrib-comments
django-cors-headers
django-debug-toolbar
requests
djoser
openai